添加swagger2,redis测试成功

+++ src/main/java/com/cer/mall/mylitemall/mbg/mapper/ColumnsPrivMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface ColumnsPrivMapper {
-    int countByExample(ColumnsPrivExample example);
+    long countByExample(ColumnsPrivExample example);
 
     int deleteByExample(ColumnsPrivExample example);
